requirements.txt
requests>=2.31.0
PyYAML>=6.0
anthropic>=0.40.0
slack-bolt>=1.19.0
pypdf>=4.0.0
google-auth>=2.35.0
